Fixing a NPE in case multi-prop definition is missing in jpos.properties
if implicitly requested by usage of the DefaultCompositeRegPopulator.
Instead throwing a runtime exception .

Restricted the check from null pointer check to empty list check,
because in case of an empty list the subsequent code may cause the NPE
seen in the tests after having applied the try-with-resources code
schema strictly.